+++ app/src/main/java/com/survivalcoding/ifkakao/data/vo/LinkListVo.kt
@@ -1,4 +1,4 @@
-package com.survivalcoding.ifkakao.domain.model
+package com.survivalcoding.ifkakao.data.vo
 
 data class LinkListVo(
     val FILE: List<FileVo>,
