+++ app/src/main/java/com/survivalcoding/ifkakao/data/vo/MoMainImageVo.kt
@@ -1,4 +1,4 @@
-package com.survivalcoding.ifkakao.domain.model
+package com.survivalcoding.ifkakao.data.vo
 
 data class MoMainImageVo(
     val contentsIdx: Int,
