+++ app/src/main/java/com/survivalcoding/ifkakao/data/vo/MoSpotlightVo.kt
@@ -1,6 +1,6 @@
-package com.survivalcoding.ifkakao.domain.model
+package com.survivalcoding.ifkakao.data.vo
 
-data class ShareImageVo(
+data class MoSpotlightVo(
     val contentsIdx: Int,
     val description: String,
     val fileSize: Int,
