+++ app/src/main/java/com/survivalcoding/ifkakao/data/vo/VideoVo.kt
@@ -1,6 +1,6 @@
-package com.survivalcoding.ifkakao.domain.model
+package com.survivalcoding.ifkakao.data.vo
 
-data class MoSpotlightVo(
+data class VideoVo(
     val contentsIdx: Int,
     val description: String,
     val fileSize: Int,
